refactor(mwl_client): migrate to mwl_adapter pattern
Introduced mwl_adapter interface to abstract MWL storage implementation,
removing 500+ lines of conditional compilation from mwl_client.

Changes:
- Created mwl_adapter interface with memory and pacs implementations
- Refactored mwl_client to use adapter pattern via dependency injection
- Removed #ifdef PACS_BRIDGE_HAS_PACS_SYSTEM blocks from mwl_client
- Updated CMakeLists.txt to include new adapter files

Implements #285
