+++ src/pacs_adapter/mwl_client.cpp
@@ -1,240 +1,70 @@
 /**
  * @file mwl_client.cpp
- * @brief Implementation of Modality Worklist client for pacs_system
+ * @brief Implementation of Modality Worklist client using mwl_adapter
  *
- * This file provides two implementations:
- * - Stub implementation (PACS_BRIDGE_STANDALONE_BUILD): In-memory storage for testing
- * - Real implementation (PACS_BRIDGE_HAS_PACS_SYSTEM): Uses pacs_system's index_database
+ * Refactored to use mwl_adapter interface which provides:
+ * - memory_mwl_adapter: In-memory storage (standalone builds)
+ * - pacs_mwl_adapter: pacs_system index_database integration
+ *
+ * This removes conditional compilation and provides consistent interface.
  */
 
 #include "pacs/bridge/pacs_adapter/mwl_client.h"
+#include "pacs/bridge/integration/mwl_adapter.h"
 #include "pacs/bridge/monitoring/bridge_metrics.h"
 #include "pacs/bridge/tracing/trace_manager.h"
 
 #include <algorithm>
 #include <atomic>
-#include <cctype>
 #include <chrono>
-#include <ctime>
 #include <mutex>
 #include <optional>
 #include <shared_mutex>
 #include <thread>
 
-#ifdef PACS_BRIDGE_HAS_PACS_SYSTEM
-#include <pacs/storage/index_database.hpp>
-#include <pacs/storage/worklist_record.hpp>
-#endif
-
 namespace pacs::bridge::pacs_adapter {
 
+// Alias for convenience
+using integration::mwl_adapter;
+using integration::mwl_adapter_error;
+using integration::mwl_query_filter;
+
 // =============================================================================
-// Conversion Utilities (for pacs_system integration)
+// Error Conversion Utilities
 // =============================================================================
 
-#ifdef PACS_BRIDGE_HAS_PACS_SYSTEM
 namespace {
 
 /**
- * @brief Convert mwl_item to pacs_system worklist_item
+ * @brief Convert mwl_adapter_error to mwl_error
  */
-[[nodiscard]] pacs::storage::worklist_item
-to_worklist_item(const mapping::mwl_item& item) {
-    pacs::storage::worklist_item wl;
+[[nodiscard]] mwl_error to_mwl_error(mwl_adapter_error adapter_error) {
+    using integration::mwl_adapter_error;
 
-    // Patient information
-    wl.patient_id = item.patient.patient_id;
-    wl.patient_name = item.patient.patient_name;
-    wl.birth_date = item.patient.patient_birth_date;
-    wl.sex = item.patient.patient_sex;
-
-    // Imaging service request
-    wl.accession_no = item.imaging_service_request.accession_number;
-
-    // Requested procedure
-    wl.requested_proc_id = item.requested_procedure.requested_procedure_id;
-    wl.study_uid = item.requested_procedure.study_instance_uid;
-    wl.referring_phys = item.requested_procedure.referring_physician_name;
-    wl.referring_phys_id = item.requested_procedure.referring_physician_id;
-
-    // Scheduled procedure step (use first step if available)
-    if (!item.scheduled_steps.empty()) {
-        const auto& sps = item.scheduled_steps[0];
-        wl.step_id = sps.scheduled_step_id;
-        wl.station_ae = sps.scheduled_station_ae_title;
-        wl.station_name = sps.scheduled_step_location;
-        wl.modality = sps.modality;
-        wl.procedure_desc = sps.scheduled_step_description;
-        wl.step_status = sps.scheduled_step_status.empty()
-                             ? "SCHEDULED"
-                             : sps.scheduled_step_status;
-
-        // Combine date and time for scheduled_datetime (YYYYMMDDHHMMSS)
-        wl.scheduled_datetime = sps.scheduled_start_date;
-        if (!sps.scheduled_start_time.empty()) {
-            wl.scheduled_datetime += sps.scheduled_start_time;
-        }
-
-        // Protocol code (serialize as JSON-like string)
-        if (!sps.protocol_code_value.empty()) {
-            wl.protocol_code = "{\"value\":\"" + sps.protocol_code_value + "\","
-                               "\"meaning\":\"" + sps.protocol_code_meaning + "\","
-                               "\"scheme\":\"" + sps.protocol_coding_scheme + "\"}";
-        }
-    } else {
-        // Generate default step_id if not provided
-        wl.step_id = item.imaging_service_request.accession_number + "_SPS1";
-        wl.step_status = "SCHEDULED";
+    switch (adapter_error) {
+        case mwl_adapter_error::not_found:
+            return mwl_error::entry_not_found;
+        case mwl_adapter_error::duplicate:
+            return mwl_error::duplicate_entry;
+        case mwl_adapter_error::invalid_data:
+            return mwl_error::invalid_data;
+        case mwl_adapter_error::query_failed:
+            return mwl_error::query_failed;
+        case mwl_adapter_error::add_failed:
+            return mwl_error::add_failed;
+        case mwl_adapter_error::update_failed:
+            return mwl_error::update_failed;
+        case mwl_adapter_error::delete_failed:
+            return mwl_error::cancel_failed;
+        case mwl_adapter_error::storage_unavailable:
+        case mwl_adapter_error::init_failed:
+            return mwl_error::connection_failed;
+        default:
+            return mwl_error::connection_failed;
     }
-
-    return wl;
-}
-
-/**
- * @brief Convert pacs_system worklist_item to mwl_item
- */
-[[nodiscard]] mapping::mwl_item
-from_worklist_item(const pacs::storage::worklist_item& wl) {
-    mapping::mwl_item item;
-
-    // Patient information
-    item.patient.patient_id = wl.patient_id;
-    item.patient.patient_name = wl.patient_name;
-    item.patient.patient_birth_date = wl.birth_date;
-    item.patient.patient_sex = wl.sex;
-
-    // Imaging service request
-    item.imaging_service_request.accession_number = wl.accession_no;
-
-    // Requested procedure
-    item.requested_procedure.requested_procedure_id = wl.requested_proc_id;
-    item.requested_procedure.study_instance_uid = wl.study_uid;
-    item.requested_procedure.referring_physician_name = wl.referring_phys;
-    item.requested_procedure.referring_physician_id = wl.referring_phys_id;
-
-    // Scheduled procedure step
-    mapping::dicom_scheduled_procedure_step sps;
-    sps.scheduled_step_id = wl.step_id;
-    sps.scheduled_station_ae_title = wl.station_ae;
-    sps.scheduled_step_location = wl.station_name;
-    sps.modality = wl.modality;
-    sps.scheduled_step_description = wl.procedure_desc;
-    sps.scheduled_step_status = wl.step_status;
-
-    // Parse scheduled_datetime (YYYYMMDDHHMMSS)
-    if (wl.scheduled_datetime.size() >= 8) {
-        sps.scheduled_start_date = wl.scheduled_datetime.substr(0, 8);
-        if (wl.scheduled_datetime.size() >= 14) {
-            sps.scheduled_start_time = wl.scheduled_datetime.substr(8, 6);
-        }
-    }
-
-    item.scheduled_steps.push_back(std::move(sps));
-
-    return item;
-}
-
-/**
- * @brief Convert mwl_query_filter to pacs_system worklist_query
- */
-[[nodiscard]] pacs::storage::worklist_query
-to_worklist_query(const mwl_query_filter& filter) {
-    pacs::storage::worklist_query query;
-
-    if (filter.patient_id && !filter.patient_id->empty()) {
-        query.patient_id = *filter.patient_id;
-    }
-    if (filter.accession_number && !filter.accession_number->empty()) {
-        query.accession_no = *filter.accession_number;
-    }
-    if (filter.patient_name && !filter.patient_name->empty()) {
-        query.patient_name = *filter.patient_name;
-    }
-    if (filter.modality && !filter.modality->empty()) {
-        query.modality = *filter.modality;
-    }
-    if (filter.scheduled_station_ae && !filter.scheduled_station_ae->empty()) {
-        query.station_ae = *filter.scheduled_station_ae;
-    }
-    if (filter.scheduled_date && !filter.scheduled_date->empty()) {
-        query.scheduled_date_from = *filter.scheduled_date;
-        query.scheduled_date_to = *filter.scheduled_date;
-    }
-    if (filter.scheduled_date_from && !filter.scheduled_date_from->empty()) {
-        query.scheduled_date_from = *filter.scheduled_date_from;
-    }
-    if (filter.scheduled_date_to && !filter.scheduled_date_to->empty()) {
-        query.scheduled_date_to = *filter.scheduled_date_to;
-    }
-
-    // Include all status or only scheduled
-    query.include_all_status = false;
-
-    if (filter.max_results > 0) {
-        query.limit = filter.max_results;
-    }
-
-    return query;
-}
-
-/**
- * @brief Parse date string (YYYYMMDD) to system_clock time_point
- *
- * Converts a DICOM-format date string to a std::chrono time_point.
- * This enables precise date-based worklist cleanup operations.
- *
- * @param date_str Date string in YYYYMMDD format
- * @return time_point if parsing succeeds, nullopt otherwise
- */
-[[nodiscard]] std::optional<std::chrono::system_clock::time_point>
-parse_date_to_timepoint(std::string_view date_str) {
-    // Validate length: YYYYMMDD (8 chars) or YYYY-MM-DD (10 chars)
-    if (date_str.length() != 8 && date_str.length() != 10) {
-        return std::nullopt;
-    }
-
-    std::tm tm = {};
-    std::string normalized(date_str);
-
-    // Handle YYYY-MM-DD format by removing dashes
-    if (date_str.length() == 10) {
-        normalized.erase(std::remove(normalized.begin(), normalized.end(), '-'),
-                        normalized.end());
-    }
-
-    if (normalized.length() != 8) {
-        return std::nullopt;
-    }
-
-    // Parse YYYYMMDD
-    try {
-        tm.tm_year = std::stoi(normalized.substr(0, 4)) - 1900;
-        tm.tm_mon = std::stoi(normalized.substr(4, 2)) - 1;
-        tm.tm_mday = std::stoi(normalized.substr(6, 2));
-        tm.tm_hour = 0;
-        tm.tm_min = 0;
-        tm.tm_sec = 0;
-        tm.tm_isdst = -1;  // Let mktime determine DST
-    } catch (const std::exception&) {
-        return std::nullopt;
-    }
-
-    // Validate parsed values
-    if (tm.tm_year < 0 || tm.tm_mon < 0 || tm.tm_mon > 11 ||
-        tm.tm_mday < 1 || tm.tm_mday > 31) {
-        return std::nullopt;
-    }
-
-    auto time_t_val = std::mktime(&tm);
-    if (time_t_val == -1) {
-        return std::nullopt;
-    }
-
-    return std::chrono::system_clock::from_time_t(time_t_val);
 }
 
 }  // namespace
-#endif  // PACS_BRIDGE_HAS_PACS_SYSTEM
 
 // =============================================================================
 // Implementation Class
@@ -242,7 +72,9 @@ parse_date_to_timepoint(std::string_view date_str) {
 
 class mwl_client::impl {
 public:
-    explicit impl(const mwl_client_config& config) : config_(config) {}
+    explicit impl(const mwl_client_config& config)
+        : config_(config),
+          adapter_(integration::create_mwl_adapter(config.pacs_database_path)) {}
 
     ~impl() { disconnect(true); }
 
@@ -259,15 +91,16 @@ public:
 
         stats_.connect_attempts++;
 
-        // Simulate connection to pacs_system
-        // In real implementation, this would establish DICOM association
-        auto result = try_connect_with_retry();
-        if (result) {
-            connected_ = true;
-            stats_.connect_successes++;
+        // Verify adapter is available
+        if (!adapter_) {
+            return std::unexpected(mwl_error::connection_failed);
         }
 
-        return result;
+        // Mark as connected (adapter handles actual storage connection)
+        connected_ = true;
+        stats_.connect_successes++;
+
+        return {};
     }
 
     void disconnect(bool graceful) {
@@ -322,51 +155,19 @@ public:
             return std::unexpected(mwl_error::connection_failed);
         }
 
-        std::unique_lock lock(mutex_);
-
-#ifdef PACS_BRIDGE_HAS_PACS_SYSTEM
-        // Real pacs_system integration: use index_database
-        if (!db_) {
-            span.set_error("Database not initialized");
-            return std::unexpected(mwl_error::connection_failed);
+        // Use adapter to add item
+        auto result = adapter_->add_item(item);
+        if (!result) {
+            auto error = to_mwl_error(result.error());
+            span.set_error(to_string(error));
+            return std::unexpected(error);
         }
 
-        // Convert mwl_item to worklist_item
-        auto wl_item = to_worklist_item(item);
-
-        // Check for duplicate by querying existing entry
-        auto existing = db_->find_worklist_item(wl_item.step_id, wl_item.accession_no);
-        if (existing.has_value()) {
-            span.set_error("Duplicate entry");
-            return std::unexpected(mwl_error::duplicate_entry);
+        // Update statistics
+        {
+            std::unique_lock lock(mutex_);
+            stats_.add_count++;
         }
-
-        // Add entry to database
-        auto add_result = db_->add_worklist_item(wl_item);
-        if (add_result.is_err()) {
-            span.set_error("Database add failed: " + add_result.error().message);
-            return std::unexpected(mwl_error::add_failed);
-        }
-
-        stats_.add_count++;
-#else
-        // Stub mode: use in-memory storage
-        // Check for duplicate
-        auto it = std::find_if(entries_.begin(), entries_.end(),
-            [&item](const mapping::mwl_item& existing) {
-                return existing.imaging_service_request.accession_number ==
-                       item.imaging_service_request.accession_number;
-            });
-
-        if (it != entries_.end()) {
-            span.set_error("Duplicate entry");
-            return std::unexpected(mwl_error::duplicate_entry);
-        }
-
-        // Add entry to in-memory storage
-        entries_.push_back(item);
-        stats_.add_count++;
-#endif
 
         // Record metrics for entry creation
         monitoring::bridge_metrics_collector::instance().record_mwl_entry_created();
@@ -407,62 +208,19 @@ public:
             return std::unexpected(mwl_error::connection_failed);
         }
 
-        std::unique_lock lock(mutex_);
-
-#ifdef PACS_BRIDGE_HAS_PACS_SYSTEM
-        // Real pacs_system integration: use index_database
-        if (!db_) {
-            span.set_error("Database not initialized");
-            return std::unexpected(mwl_error::connection_failed);
+        // Use adapter to update item
+        auto result = adapter_->update_item(accession_number, item);
+        if (!result) {
+            auto error = to_mwl_error(result.error());
+            span.set_error(to_string(error));
+            return std::unexpected(error);
         }
 
-        // Find existing entry by accession number using query
-        pacs::storage::worklist_query query;
-        query.accession_no = std::string(accession_number);
-        query.include_all_status = true;
-
-        auto query_result = db_->query_worklist(query);
-        if (query_result.is_err() || query_result.value().empty()) {
-            span.set_error("Entry not found");
-            return std::unexpected(mwl_error::entry_not_found);
+        // Update statistics
+        {
+            std::unique_lock lock(mutex_);
+            stats_.update_count++;
         }
-
-        const auto& existing_wl = query_result.value()[0];
-
-        // Convert updated item to worklist_item and merge
-        auto wl_item = to_worklist_item(item);
-
-        // Update status if the new status is provided
-        std::string new_status = wl_item.step_status.empty()
-                                     ? existing_wl.step_status
-                                     : wl_item.step_status;
-
-        auto update_result = db_->update_worklist_status(
-            existing_wl.step_id, std::string(accession_number), new_status);
-
-        if (update_result.is_err()) {
-            span.set_error("Database update failed: " + update_result.error().message);
-            return std::unexpected(mwl_error::update_failed);
-        }
-
-        stats_.update_count++;
-#else
-        // Stub mode: use in-memory storage
-        auto it = std::find_if(entries_.begin(), entries_.end(),
-            [accession_number](const mapping::mwl_item& existing) {
-                return existing.imaging_service_request.accession_number ==
-                       accession_number;
-            });
-
-        if (it == entries_.end()) {
-            span.set_error("Entry not found");
-            return std::unexpected(mwl_error::entry_not_found);
-        }
-
-        // Update entry with non-empty fields from item
-        update_mwl_item(*it, item);
-        stats_.update_count++;
-#endif
 
         // Record metrics for entry update
         monitoring::bridge_metrics_collector::instance().record_mwl_entry_updated();
@@ -502,54 +260,19 @@ public:
             return std::unexpected(mwl_error::connection_failed);
         }
 
-        std::unique_lock lock(mutex_);
-
-#ifdef PACS_BRIDGE_HAS_PACS_SYSTEM
-        // Real pacs_system integration: use index_database
-        if (!db_) {
-            span.set_error("Database not initialized");
-            return std::unexpected(mwl_error::connection_failed);
+        // Use adapter to delete item
+        auto result = adapter_->delete_item(accession_number);
+        if (!result) {
+            auto error = to_mwl_error(result.error());
+            span.set_error(to_string(error));
+            return std::unexpected(error);
         }
 
-        // Find entry by accession number to get step_id
-        pacs::storage::worklist_query query;
-        query.accession_no = std::string(accession_number);
-        query.include_all_status = true;
-
-        auto query_result = db_->query_worklist(query);
-        if (query_result.is_err() || query_result.value().empty()) {
-            span.set_error("Entry not found");
-            return std::unexpected(mwl_error::entry_not_found);
+        // Update statistics
+        {
+            std::unique_lock lock(mutex_);
+            stats_.cancel_count++;
         }
-
-        const auto& existing_wl = query_result.value()[0];
-
-        // Delete entry from database
-        auto delete_result = db_->delete_worklist_item(
-            existing_wl.step_id, std::string(accession_number));
-
-        if (delete_result.is_err()) {
-            span.set_error("Database delete failed: " + delete_result.error().message);
-            return std::unexpected(mwl_error::cancel_failed);
-        }
-
-        stats_.cancel_count++;
-#else
-        // Stub mode: use in-memory storage
-        auto it = std::find_if(entries_.begin(), entries_.end(),
-            [accession_number](const mapping::mwl_item& existing) {
-                return existing.imaging_service_request.accession_number ==
-                       accession_number;
-            });
-
-        if (it == entries_.end()) {
-            span.set_error("Entry not found");
-            return std::unexpected(mwl_error::entry_not_found);
-        }
-
-        entries_.erase(it);
-        stats_.cancel_count++;
-#endif
 
         // Record metrics for entry cancellation
         monitoring::bridge_metrics_collector::instance().record_mwl_entry_cancelled();
@@ -583,47 +306,19 @@ public:
             return std::unexpected(mwl_error::connection_failed);
         }
 
-        std::shared_lock lock(mutex_);
-
-        std::vector<mapping::mwl_item> results;
-
-#ifdef PACS_BRIDGE_HAS_PACS_SYSTEM
-        // Real pacs_system integration: use index_database
-        if (!db_) {
-            span.set_error("Database not initialized");
-            return std::unexpected(mwl_error::connection_failed);
+        // Use adapter to query items
+        auto result = adapter_->query_items(filter);
+        if (!result) {
+            auto error = to_mwl_error(result.error());
+            span.set_error(to_string(error));
+            return std::unexpected(error);
         }
 
-        // Convert filter to pacs_system worklist_query
-        auto wl_query = to_worklist_query(filter);
-
-        auto query_result = db_->query_worklist(wl_query);
-        if (query_result.is_err()) {
-            span.set_error("Database query failed: " + query_result.error().message);
-            return std::unexpected(mwl_error::query_failed);
+        // Update statistics
+        {
+            std::unique_lock lock(mutex_);
+            stats_.query_count++;
         }
-
-        // Convert worklist_items to mwl_items
-        for (const auto& wl_item : query_result.value()) {
-            results.push_back(from_worklist_item(wl_item));
-        }
-
-        stats_.query_count++;
-#else
-        // Stub mode: use in-memory storage
-        for (const auto& entry : entries_) {
-            if (matches_filter(entry, filter)) {
-                results.push_back(entry);
-
-                if (filter.max_results > 0 &&
-                    results.size() >= filter.max_results) {
-                    break;
-                }
-            }
-        }
-
-        stats_.query_count++;
-#endif
 
         auto elapsed = std::chrono::duration_cast<std::chrono::milliseconds>(
             std::chrono::steady_clock::now() - start_time);
@@ -637,11 +332,11 @@ public:
         update_avg_operation_time(elapsed.count());
 
         span.set_attribute("pacs.duration_ms", static_cast<int64_t>(elapsed.count()));
-        span.set_attribute("pacs.result_count", static_cast<int64_t>(results.size()));
+        span.set_attribute("pacs.result_count", static_cast<int64_t>(result->size()));
         span.set_attribute("pacs.success", true);
 
         return mwl_client::query_result{
-            .items = std::move(results),
+            .items = std::move(*result),
             .elapsed_time = elapsed,
             .has_more = false,
             .total_count = std::nullopt
@@ -688,29 +383,7 @@ public:
             return false;
         }
 
-        std::shared_lock lock(mutex_);
-
-#ifdef PACS_BRIDGE_HAS_PACS_SYSTEM
-        // Real pacs_system integration: use index_database
-        if (!db_) {
-            return false;
-        }
-
-        pacs::storage::worklist_query query;
-        query.accession_no = std::string(accession_number);
-        query.include_all_status = true;
-        query.limit = 1;
-
-        auto query_result = db_->query_worklist(query);
-        return query_result.is_ok() && !query_result.value().empty();
-#else
-        // Stub mode: use in-memory storage
-        return std::any_of(entries_.begin(), entries_.end(),
-            [accession_number](const mapping::mwl_item& entry) {
-                return entry.imaging_service_request.accession_number ==
-                       accession_number;
-            });
-#endif
+        return adapter_->exists(accession_number);
     }
 
     std::expected<mapping::mwl_item, mwl_error>
@@ -723,39 +396,12 @@ public:
             return std::unexpected(mwl_error::connection_failed);
         }
 
-        std::shared_lock lock(mutex_);
-
-#ifdef PACS_BRIDGE_HAS_PACS_SYSTEM
-        // Real pacs_system integration: use index_database
-        if (!db_) {
-            return std::unexpected(mwl_error::connection_failed);
+        auto result = adapter_->get_item(accession_number);
+        if (!result) {
+            return std::unexpected(to_mwl_error(result.error()));
         }
 
-        pacs::storage::worklist_query query;
-        query.accession_no = std::string(accession_number);
-        query.include_all_status = true;
-        query.limit = 1;
-
-        auto query_result = db_->query_worklist(query);
-        if (query_result.is_err() || query_result.value().empty()) {
-            return std::unexpected(mwl_error::entry_not_found);
-        }
-
-        return from_worklist_item(query_result.value()[0]);
-#else
-        // Stub mode: use in-memory storage
-        auto it = std::find_if(entries_.begin(), entries_.end(),
-            [accession_number](const mapping::mwl_item& entry) {
-                return entry.imaging_service_request.accession_number ==
-                       accession_number;
-            });
-
-        if (it == entries_.end()) {
-            return std::unexpected(mwl_error::entry_not_found);
-        }
-
-        return *it;
-#endif
+        return *result;
     }
 
     // =========================================================================
@@ -791,78 +437,23 @@ public:
             return std::unexpected(mwl_error::connection_failed);
         }
 
-        std::unique_lock lock(mutex_);
-
-        size_t cancelled = 0;
-
-#ifdef PACS_BRIDGE_HAS_PACS_SYSTEM
-        // Real pacs_system integration: use index_database
-        if (!db_) {
-            return std::unexpected(mwl_error::connection_failed);
+        // Use adapter to delete items before date
+        auto result = adapter_->delete_items_before(before_date);
+        if (!result) {
+            return std::unexpected(to_mwl_error(result.error()));
         }
 
-        // Parse date string (YYYYMMDD format) to time_point
-        // This provides exact date-based cleanup without timezone ambiguities
-        auto before_time = parse_date_to_timepoint(before_date);
-        if (!before_time) {
-            return std::unexpected(mwl_error::invalid_data);
-        }
-
-        // Use pacs_system's date-based cleanup API for precise cleanup
-        auto cleanup_result = db_->cleanup_worklist_items_before(*before_time);
-        if (cleanup_result.is_err()) {
-            return std::unexpected(mwl_error::cancel_failed);
-        }
-
-        cancelled = cleanup_result.value();
+        size_t cancelled = *result;
 
         // Update statistics for each cancelled entry
-        for (size_t i = 0; i < cancelled; ++i) {
-            stats_.cancel_count++;
-            monitoring::bridge_metrics_collector::instance()
-                .record_mwl_entry_cancelled();
-        }
-#else
-        // Stub mode: use in-memory storage
-        // Validate date format (YYYYMMDD - 8 chars)
-        if (before_date.length() != 8) {
-            return std::unexpected(mwl_error::invalid_data);
-        }
-
-        // Check if all characters are digits
-        for (char c : before_date) {
-            if (!std::isdigit(static_cast<unsigned char>(c))) {
-                return std::unexpected(mwl_error::invalid_data);
-            }
-        }
-
-        std::string before_date_str(before_date);
-
-        auto it = entries_.begin();
-        while (it != entries_.end()) {
-            bool should_cancel = false;
-
-            if (!it->scheduled_steps.empty()) {
-                const auto& sps = it->scheduled_steps[0];
-                if (!sps.scheduled_start_date.empty() &&
-                    sps.scheduled_start_date < before_date_str) {
-                    should_cancel = true;
-                }
-            }
-
-            if (should_cancel) {
-                it = entries_.erase(it);
-                cancelled++;
+        {
+            std::unique_lock lock(mutex_);
+            for (size_t i = 0; i < cancelled; ++i) {
                 stats_.cancel_count++;
-
-                // Record metrics for each entry cancellation
                 monitoring::bridge_metrics_collector::instance()
                     .record_mwl_entry_cancelled();
-            } else {
-                ++it;
             }
         }
-#endif
 
         return cancelled;
     }
@@ -888,43 +479,6 @@ private:
     // Helper Methods
     // =========================================================================
 
-    std::expected<void, mwl_error> try_connect_with_retry() {
-        for (size_t attempt = 0; attempt <= config_.max_retries; ++attempt) {
-            if (attempt > 0) {
-                std::this_thread::sleep_for(config_.retry_delay);
-            }
-
-#ifdef PACS_BRIDGE_HAS_PACS_SYSTEM
-            // Real pacs_system integration: open index_database
-            if (!db_) {
-                auto db_result = pacs::storage::index_database::open(
-                    config_.pacs_database_path);
-
-                if (db_result.is_err()) {
-                    // Log error and retry
-                    continue;
-                }
-
-                db_ = std::move(db_result.value());
-            }
-
-            // Verify database is open and accessible
-            if (db_ && db_->is_open()) {
-                return {};
-            }
-#else
-            // Stub mode: Connection always succeeds (in-memory storage)
-            // In real implementation:
-            // - Create DICOM association
-            // - Negotiate presentation contexts for MWL
-            // - Verify AE titles
-            return {};
-#endif
-        }
-
-        return std::unexpected(mwl_error::connection_failed);
-    }
-
     bool ensure_connected() {
         if (is_connected()) {
             return true;
@@ -933,194 +487,6 @@ private:
         auto result = connect();
         return result.has_value();
     }
-
-#ifndef PACS_BRIDGE_HAS_PACS_SYSTEM
-    // Stub mode helper methods
-    bool matches_filter(const mapping::mwl_item& entry,
-                       const mwl_query_filter& filter) const {
-        // Patient ID filter
-        if (filter.patient_id && !filter.patient_id->empty()) {
-            if (entry.patient.patient_id != *filter.patient_id) {
-                return false;
-            }
-        }
-
-        // Accession number filter
-        if (filter.accession_number && !filter.accession_number->empty()) {
-            if (entry.imaging_service_request.accession_number !=
-                *filter.accession_number) {
-                return false;
-            }
-        }
-
-        // Patient name filter (supports wildcards with *)
-        if (filter.patient_name && !filter.patient_name->empty()) {
-            if (!matches_wildcard(entry.patient.patient_name,
-                                  *filter.patient_name)) {
-                return false;
-            }
-        }
-
-        // Referring physician filter
-        if (filter.referring_physician && !filter.referring_physician->empty()) {
-            if (!matches_wildcard(
-                    entry.requested_procedure.referring_physician_name,
-                    *filter.referring_physician)) {
-                return false;
-            }
-        }
-
-        // Scheduled procedure step filters
-        if (!entry.scheduled_steps.empty()) {
-            const auto& sps = entry.scheduled_steps[0];
-
-            // Modality filter
-            if (filter.modality && !filter.modality->empty()) {
-                if (sps.modality != *filter.modality) {
-                    return false;
-                }
-            }
-
-            // Scheduled station AE filter
-            if (filter.scheduled_station_ae &&
-                !filter.scheduled_station_ae->empty()) {
-                if (sps.scheduled_station_ae_title !=
-                    *filter.scheduled_station_ae) {
-                    return false;
-                }
-            }
-
-            // Scheduled date filter (exact match)
-            if (filter.scheduled_date && !filter.scheduled_date->empty()) {
-                if (sps.scheduled_start_date != *filter.scheduled_date) {
-                    return false;
-                }
-            }
-
-            // Scheduled date range filter
-            if (filter.scheduled_date_from &&
-                !filter.scheduled_date_from->empty()) {
-                if (sps.scheduled_start_date < *filter.scheduled_date_from) {
-                    return false;
-                }
-            }
-
-            if (filter.scheduled_date_to && !filter.scheduled_date_to->empty()) {
-                if (sps.scheduled_start_date > *filter.scheduled_date_to) {
-                    return false;
-                }
-            }
-
-            // SPS status filter
-            if (filter.sps_status && !filter.sps_status->empty()) {
-                if (sps.scheduled_step_status != *filter.sps_status) {
-                    return false;
-                }
-            }
-        }
-
-        return true;
-    }
-
-    bool matches_wildcard(std::string_view text,
-                         std::string_view pattern) const {
-        // Simple wildcard matching with * as wildcard
-        if (pattern.empty()) {
-            return true;
-        }
-
-        if (pattern == "*") {
-            return true;
-        }
-
-        // Check for prefix match (pattern ends with *)
-        if (pattern.back() == '*') {
-            auto prefix = pattern.substr(0, pattern.size() - 1);
-            return text.substr(0, prefix.size()) == prefix;
-        }
-
-        // Check for suffix match (pattern starts with *)
-        if (pattern.front() == '*') {
-            auto suffix = pattern.substr(1);
-            if (text.size() < suffix.size()) {
-                return false;
-            }
-            return text.substr(text.size() - suffix.size()) == suffix;
-        }
-
-        // Exact match
-        return text == pattern;
-    }
-
-    void update_mwl_item(mapping::mwl_item& existing,
-                        const mapping::mwl_item& updates) {
-        // Update patient fields if non-empty
-        if (!updates.patient.patient_id.empty()) {
-            existing.patient.patient_id = updates.patient.patient_id;
-        }
-        if (!updates.patient.patient_name.empty()) {
-            existing.patient.patient_name = updates.patient.patient_name;
-        }
-        if (!updates.patient.patient_birth_date.empty()) {
-            existing.patient.patient_birth_date =
-                updates.patient.patient_birth_date;
-        }
-        if (!updates.patient.patient_sex.empty()) {
-            existing.patient.patient_sex = updates.patient.patient_sex;
-        }
-
-        // Update imaging service request fields
-        if (!updates.imaging_service_request.requesting_physician.empty()) {
-            existing.imaging_service_request.requesting_physician =
-                updates.imaging_service_request.requesting_physician;
-        }
-        if (!updates.imaging_service_request.requesting_service.empty()) {
-            existing.imaging_service_request.requesting_service =
-                updates.imaging_service_request.requesting_service;
-        }
-
-        // Update requested procedure fields
-        if (!updates.requested_procedure.requested_procedure_description
-                 .empty()) {
-            existing.requested_procedure.requested_procedure_description =
-                updates.requested_procedure.requested_procedure_description;
-        }
-        if (!updates.requested_procedure.referring_physician_name.empty()) {
-            existing.requested_procedure.referring_physician_name =
-                updates.requested_procedure.referring_physician_name;
-        }
-
-        // Update scheduled procedure steps
-        if (!updates.scheduled_steps.empty()) {
-            if (existing.scheduled_steps.empty()) {
-                existing.scheduled_steps = updates.scheduled_steps;
-            } else {
-                auto& existing_sps = existing.scheduled_steps[0];
-                const auto& update_sps = updates.scheduled_steps[0];
-
-                if (!update_sps.scheduled_start_date.empty()) {
-                    existing_sps.scheduled_start_date =
-                        update_sps.scheduled_start_date;
-                }
-                if (!update_sps.scheduled_start_time.empty()) {
-                    existing_sps.scheduled_start_time =
-                        update_sps.scheduled_start_time;
-                }
-                if (!update_sps.modality.empty()) {
-                    existing_sps.modality = update_sps.modality;
-                }
-                if (!update_sps.scheduled_station_ae_title.empty()) {
-                    existing_sps.scheduled_station_ae_title =
-                        update_sps.scheduled_station_ae_title;
-                }
-                if (!update_sps.scheduled_performing_physician.empty()) {
-                    existing_sps.scheduled_performing_physician =
-                        update_sps.scheduled_performing_physician;
-                }
-            }
-        }
-    }
-#endif  // !PACS_BRIDGE_HAS_PACS_SYSTEM
 
     void update_avg_operation_time(double elapsed_ms) {
         // Running average calculation
@@ -1139,17 +505,10 @@ private:
     // =========================================================================
 
     mwl_client_config config_;
+    std::shared_ptr<mwl_adapter> adapter_;
     mutable std::shared_mutex mutex_;
     std::atomic<bool> connected_{false};
     mwl_client::statistics stats_;
-
-#ifdef PACS_BRIDGE_HAS_PACS_SYSTEM
-    // Real pacs_system integration: use index_database for persistent storage
-    std::unique_ptr<pacs::storage::index_database> db_;
-#else
-    // Stub mode: use in-memory storage
-    std::vector<mapping::mwl_item> entries_;
-#endif
 };
 
 // =============================================================================
